+++ back/engines/commercial/idea_custom_fields/config/routes.rb
@@ -8,6 +8,7 @@ IdeaCustomFields::Engine.routes.draw do
           resources :custom_fields, only: %i[index show], controller: 'idea_custom_fields' do
             patch 'by_code/:code', action: 'upsert_by_code', on: :collection
             patch 'update/:id', action: 'update', on: :collection
+            patch 'update_all', on: :collection
           end
         end
       end