+++ front/app/modules/commercial/customizable_navbar/admin/containers/EditPageForm/EditPageFormNavbar.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import { withRouter, WithRouterProps } from 'utils/cl-router/withRouter';
-import { Formik, FormikProps } from 'formik';
 
 // components
 import PageFormWithNavbarNameField, {
-  validatePageForm,
   FormValues,
 } from '../../components/PageFormWithNavbarNameField';
 import SectionFormWrapper from 'containers/Admin/pagesAndMenu/components/SectionFormWrapper';
@@ -12,7 +10,7 @@ import { pagesAndMenuBreadcrumb } from 'containers/Admin/pagesAndMenu/breadcrumb
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
-import { getInitialFormValues, createPageUpdateData } from './utils';
+import { createPageUpdateData, getInitialFormValues } from './utils';
 
 // i18n
 import { injectIntl } from 'utils/cl-intl';
@@ -26,7 +24,6 @@ import { handleAddPageFiles, handleRemovePageFiles } from 'services/pageFiles';
 import useAppConfigurationLocales from 'hooks/useAppConfigurationLocales';
 import useRemoteFiles from 'hooks/useRemoteFiles';
 import usePage from 'hooks/usePage';
-import usePageSlugs from 'hooks/usePageSlugs';
 import useLocalize from 'hooks/useLocalize';
 
 const EditPageFormNavbar = ({
@@ -41,54 +38,34 @@ const EditPageFormNavbar = ({
     resourceType: 'page',
     resourceId: !isNilOrError(page) ? page.id : null,
   });
-  const pageSlugs = usePageSlugs();
 
-  if (
-    isNilOrError(page) ||
-    isNilOrError(appConfigurationLocales) ||
-    isNilOrError(pageSlugs)
-  ) {
+  if (isNilOrError(page) || isNilOrError(appConfigurationLocales)) {
     return null;
   }
 
-  const handleSubmit = async (
-    values: FormValues,
-    { setSubmitting, setStatus }
-  ) => {
+  const handleSubmit = async (values: FormValues) => {
     const localPageFiles = values.local_page_files;
 
-    try {
-      const promises: Promise<any>[] = [
-        updatePage(pageId, createPageUpdateData(page, values)),
-      ];
+    const promises: Promise<any>[] = [
+      updatePage(pageId, createPageUpdateData(page, values)),
+    ];
 
-      if (!isNilOrError(localPageFiles)) {
-        const addPromise = handleAddPageFiles(
-          pageId,
-          localPageFiles,
-          remotePageFiles
-        );
-        const removePromise = handleRemovePageFiles(
-          pageId,
-          localPageFiles,
-          remotePageFiles
-        );
+    if (!isNilOrError(localPageFiles)) {
+      const addPromise = handleAddPageFiles(
+        pageId,
+        localPageFiles,
+        remotePageFiles
+      );
+      const removePromise = handleRemovePageFiles(
+        pageId,
+        localPageFiles,
+        remotePageFiles
+      );
 
-        promises.push(addPromise, removePromise);
-      }
-
-      await Promise.all(promises);
-
-      setStatus('success');
-      setSubmitting(false);
-    } catch (error) {
-      setStatus('error');
-      setSubmitting(false);
+      promises.push(addPromise, removePromise);
     }
-  };
 
-  const renderFn = (props: FormikProps<FormValues>) => {
-    return <PageFormWithNavbarNameField {...props} pageId={pageId} />;
+    await Promise.all(promises);
   };
 
   return (
@@ -104,17 +81,10 @@ const EditPageFormNavbar = ({
         },
       ]}
     >
-      <Formik
-        initialValues={getInitialFormValues(page, remotePageFiles)}
+      <PageFormWithNavbarNameField
+        pageId={pageId}
         onSubmit={handleSubmit}
-        render={renderFn}
-        validate={validatePageForm(
-          appConfigurationLocales,
-          pageSlugs,
-          page.attributes.slug
-        )}
-        validateOnChange={false}
-        validateOnBlur={false}
+        defaultValues={getInitialFormValues(page, remotePageFiles)}
       />
     </SectionFormWrapper>
   );